+++ CHANGELOG.md
@@ -6,3 +6,4 @@ HEAD
 
 * first commit
 * Add platformio library.json
+* Add first MQTTClient class and include ArduinoJson lib dep